+++ pkg/providers/appsec/testdata/TestDSConfiguration/match_by_id.tf
@@ -19,3 +19,7 @@ output "configsedgeconfiglist" {
   value = data.akamai_appsec_configuration.test.output_text
 }
 
+output "host_names" {
+  value = data.akamai_appsec_configuration.test.host_names
+}
+